+++ src/frontend/src/pages/NewCalendarPage/CalendarTab.tsx
@@ -5,17 +5,28 @@ import { Box } from '@mui/material';
 import FullPageTabs from '../../components/FullPageTabs';
 import { useState } from 'react';
 import { useCurrentUser } from '../../hooks/users.hooks';
-import { ConflictStatus, isHead, isLead } from 'shared';
-import { useAllCalendars, useAllEventTypes, useFilterEvents } from '../../hooks/calendar.hooks';
+import { Event, ConflictStatus, isHead, isLead } from 'shared';
+import {
+  EditEventArgs,
+  useAllCalendars,
+  useAllEventTypes,
+  useFilterEvents,
+  useUploadManyDocuments
+} from '../../hooks/calendar.hooks';
 import LoadingIndicator from '../../components/LoadingIndicator';
 import ErrorPage from '../ErrorPage';
 import { filterEventTransformer } from '../../apis/transformers/calendar.transformer';
 import EventsTable from './EventsTable';
+import { EventRoutePayload } from './Components/EventModal';
+import { useToast } from '../../hooks/toasts.hooks';
 
 const CalendarTab: React.FC = () => {
   const [tabIndex, setTabIndex] = useState<number>(0);
   const user = useCurrentUser();
   const canViewReviews = isHead(user.role) || isLead(user.role);
+  const toast = useToast();
+  const { mutateAsync: uploadDocuments } = useUploadManyDocuments();
+
   const tabs = [
     { tabUrlValue: 'mainCalendar', tabName: 'Calendar' },
     { tabUrlValue: 'yourEvents', tabName: 'Your Events' }
@@ -82,6 +93,60 @@ const CalendarTab: React.FC = () => {
 
   if (canViewReviews) tabs.push({ tabUrlValue: 'reviews', tabName: 'Review Bookings' });
 
+  const handleEditSubmit = async (
+    data: EventRoutePayload,
+    event: Event,
+    editEvent: (editArgs: EditEventArgs) => Promise<Event>,
+    onClose: () => void
+  ) => {
+    if (!event) return;
+
+    try {
+      const { scheduleSlot, documentFiles, ...eventData } = data;
+
+      if (!scheduleSlot || scheduleSlot.length === 0) {
+        throw new Error('Missing scheduleSlot');
+      }
+
+      // Convert EventRoutePayload to EditEventArgs format
+      const editArgs: EditEventArgs = {
+        ...eventData,
+        status: event.status, // Use existing event status
+        documents: event.documents.map((doc) => ({
+          name: doc.name,
+          googleFileId: doc.googleFileId
+        })),
+        scheduleSlot: scheduleSlot.map((slot) => ({
+          days: slot.days,
+          startTime: slot.startTime,
+          endTime: slot.endTime,
+          recurrenceNumber: slot.recurrenceNumber,
+          initialDateScheduled: slot.initialDateScheduled,
+          allDay: slot.allDay
+        }))
+      };
+
+      const editedEvent = await editEvent(editArgs);
+
+      // Upload new documents if any
+      const filesToUpload = documentFiles.map((doc) => doc.file).filter((file): file is File => file !== undefined);
+
+      if (filesToUpload.length > 0) {
+        await uploadDocuments({
+          id: editedEvent.eventId,
+          files: filesToUpload
+        });
+      }
+
+      toast.success('Event updated successfully!');
+      onClose();
+    } catch (err) {
+      if (err instanceof Error) {
+        toast.error(err.message);
+      }
+    }
+  };
+
   return (
     <PageLayout
       title="Calendar"
@@ -112,6 +177,7 @@ const CalendarTab: React.FC = () => {
           reviewEvents={reviewEvents ?? []}
           allEventTypes={allEventTypes}
           allCalendars={allCalendars}
+          handleEditSubmit={handleEditSubmit}
         />
       )}
     </PageLayout>
